cambiato il tipo di leaderboardresponse, aggiunto un metodo di avatarrepository, fix di leaderboardservice per la modifica dei dto

+++ src/main/java/it/project_work/app_arcade/dto/LeaderboardResponse.java
@@ -3,8 +3,8 @@ package it.project_work.app_arcade.dto;
 public record LeaderboardResponse(
         String username,
         String avatarUrl,
-        Long totalScore, // somma bestScore di tutti i giochi
-        Long totalPlayed, // somma played_count di tutti i giochi
+        Integer totalScore, // somma bestScore di tutti i giochi
+        Integer totalPlayed, // somma played_count di tutti i giochi
         Integer level
         ) {
